requirements.txt
Flask
boto3
python-dotenv
mysql-connector-python
